pyproject.toml
[project]
name = "fap"
version = "0.1.0"
description = "API for AST"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[standard]>=0.116.1",
    "faster-whisper>=1.2.1",
    "google-generativeai>=0.3.0",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
]

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
